# Use `torch.repeat_interleave()` to generate repeated indices faster