--- a/timm/data/distributed_sampler.py
+++ b/timm/data/distributed_sampler.py
@@ -104,12 +104,12 @@ class RepeatAugSampler(Sampler):
         g = torch.Generator()
         g.manual_seed(self.epoch)
         if self.shuffle:
-            indices = torch.randperm(len(self.dataset), generator=g).tolist()
+            indices = torch.randperm(len(self.dataset), generator=g)
         else:
-            indices = list(range(len(self.dataset)))
+            indices = torch.range(start=0, end=len(self.dataset))
 
         # produce repeats e.g. [0, 0, 0, 1, 1, 1, 2, 2, 2....]
-        indices = np.repeat(indices, repeats=self.num_repeats, axis=0)
+        indices = torch.repeat_interleave(indices, repeats=self.num_repeats, dim=0)
         # add extra samples to make it evenly divisible
         padding_size = self.total_size - len(indices)
         indices += indices[:padding_size]