--- a/timm/data/distributed_sampler.py
+++ b/timm/data/distributed_sampler.py
@@ -1,6 +1,5 @@
 import math
 import torch
-import numpy as np
 from torch.utils.data import Sampler
 import torch.distributed as dist