@MichaelMonashev 2021-12-29
Did you compare speed with torch.repeat_interleave() https://pytorch.org/docs/stable/generated/torch.repeat_interleave.html ?