@kozistr 2021-12-29
> Did you compare speed with torch.repeat_interleave() https://pytorch.org/docs/stable/generated/torch.repeat_interleave.html ?

Thanks for your great idea! I thought casting List to Tensor and Tensor to Numpy could be expensive, but the method you suggested was faster : )

I just updated the benchmark code with `torch.repeat_interleave()` benchmark.