@kozistr 2022-01-02
> @kozistr thanks, re the final conversion back to list, did you check if doing to list and adding extra samples is faster than
> 
> ```
> padding_size = self.total_size - len(indices)
> if padding_size > 0:
>   indices = torch.cat([indices, indices[:padding_size]])
> ```
> 
> Probably not a huge difference either way but for purity of impl if rest is in tensor form now.... plus the memory occupied by tensor probably quite a bit better than a list, especially for a really large dataset

I agree with you. The cost of casting is somewhat expensive and using tensor could be more efficient I guess too!